+++ core/src/main/kotlin/net/onedaybeard/graftt/ResultKt.kt
@@ -19,27 +19,27 @@ fun <T> resultOf(f: () -> T): Result<T, Msg> {
 fun <T> resultOf(vararg require: Result<*, Msg>, f: () -> T): Result<T, Msg> {
     return when (val e = require.find { it is Err<Msg> }) {
         null -> resultOf(f)
-        else -> Err(e.unwrapError()).andThen { resultOf(f) }
+        else -> Err(e.unwrapError())
     }
 }
 
-@Suppress("UNCHECKED_CAST")
 /** Transforms [Msg] unless it is [Msg.Error] */
-fun <T, U> Result<T, Msg>.mapSafeError(f: (Msg) -> U): Result<T, U> {
+fun <T> Result<T, Msg>.mapSafeError(f: (Msg) -> Msg): Result<T, Msg> {
     return when (val e = getError()) {
         null         -> this
         is Msg.Error -> this
         else         -> Err(f(e))
-    } as Result<T, U>
+    }
 }
 
 @Suppress("UNCHECKED_CAST")
-/** Transforms [Msg] unless it is [Msg.Error] */
+/** Transforms [Msg] unless it is [Msg.None] */
 fun <T> Result<T, Msg>.safeRecover(f: (Msg) -> T): Result<T, Msg> {
     return when (val e = getError()) {
-        null        -> this
-        is Msg.None -> Ok(f(e))
-        else        -> this
+        null                    -> this
+        is Msg.None             -> Ok(f(e))
+        is Msg.NoSuchAnnotation -> Ok(f(e))
+        else                    -> this
     }
 }
 
@@ -47,7 +47,7 @@ fun <T> Result<T, Msg>.safeRecover(f: (Msg) -> T): Result<T, Msg> {
 fun <V, E> combine(vararg results: () -> Result<V, E>): Result<List<V>, E> {
     return Ok(results.map { f ->
         when (val result = f()) {
-            is Ok -> result.value
+            is Ok  -> result.value
             is Err -> return result
         }
     })