+++ graftt-maven-plugin/src/main/kotlin/net/onedaybeard/graftt/GrafttMojo.kt
@@ -1,6 +1,9 @@
 package net.onedaybeard.graftt
 
 import com.github.michaelbull.result.*
+import net.onedaybeard.graftt.asm.classNode
+import net.onedaybeard.graftt.asm.shortName
+import net.onedaybeard.graftt.asm.toBytes
 import net.onedaybeard.graftt.graft.isTransplant
 import net.onedaybeard.graftt.graft.transplant
 import net.onedaybeard.graftt.graft.readRecipientType
@@ -62,7 +65,7 @@ class GrafttMojo : AbstractMojo() {
         // commence surgery
         transplants
             .onEach { cn -> transplant(cn, remapper) }
-            .let(this::logSummary)
+            .let(::logSummary)
 
         if (!keepTransplants) {
             classNodes(classDir)
@@ -91,35 +94,9 @@ class GrafttMojo : AbstractMojo() {
         return transplants.filter(ClassNode::isTransplant)
     }
 
-    private fun logSummary(donated: List<ClassNode>) {
-        fun format(kv: Pair<String, String>, delim: Char = '.'): String {
-            return "$delim".repeat(max(2, LINE_WIDTH - 2 - kv.length))
-                .let { "${kv.first} $it ${kv.second}" }
-        }
-
-        fun header(header: String, delim: Char = ' '): String {
-            return "$delim".repeat(max(2, LINE_WIDTH - 2 - header.length))
-                .let { "$header: $it" }
-        }
-
-        log.info(format("graftt surgical summary:" to "${donated.size}", ' '))
-        log.info("-".repeat(LINE_WIDTH))
-        log.info(header("CONFIG"))
-        log.info(format("enable" to "$enable"))
-        log.info(format("classDir" to "$classDir"))
-        log.info(format("keepTransplants" to "$keepTransplants"))
-        log.info("-".repeat(LINE_WIDTH))
-        log.info(header("TRANSPLANTS"))
-        donated
-            .map { cn -> cn to readRecipientType(cn).andThen(this::loadClassNode).unwrap() }
-            .map { (d, r) -> format(d.shortName to r.shortName) }
-            .forEach(log::info)
-        log.info("-".repeat(LINE_WIDTH))
-    }
-
     private fun transplant(donor: ClassNode, remapper: Remapper) {
         transplant(donor, this::loadClassNode, remapper)
-            .andThen(this::save)
+            .andThen(::save)
             .onFailure(`(╯°□°）╯︵ ┻━┻`)
     }
 
@@ -144,5 +121,31 @@ class GrafttMojo : AbstractMojo() {
 
         return project.artifacts.find { it matching this }
             ?: throw IllegalStateException("unable to resolve dependency: $artifactId")
+    }
+
+    private fun logSummary(donated: List<ClassNode>) {
+        fun format(kv: Pair<String, String>, delim: Char = '.'): String {
+            return "$delim".repeat(max(2, LINE_WIDTH - 2 - kv.length))
+                .let { "${kv.first} $it ${kv.second}" }
+        }
+
+        fun header(header: String, delim: Char = ' '): String {
+            return "$delim".repeat(max(2, LINE_WIDTH - 2 - header.length))
+                .let { "$header: $it" }
+        }
+
+        log.info(format("graftt surgical summary:" to "${donated.size}", ' '))
+        log.info("-".repeat(LINE_WIDTH))
+        log.info(header("CONFIG"))
+        log.info(format("enable" to "$enable"))
+        log.info(format("classDir" to "$classDir"))
+        log.info(format("keepTransplants" to "$keepTransplants"))
+        log.info("-".repeat(LINE_WIDTH))
+        log.info(header("TRANSPLANTS"))
+        donated
+            .map { cn -> cn to readRecipientType(cn).andThen(this::loadClassNode).unwrap() }
+            .map { (d, r) -> format(d.shortName to r.shortName) }
+            .forEach(log::info)
+        log.info("-".repeat(LINE_WIDTH))
     }
 }